-Anonymity Bar resized slightly (like you suggested)
-"Settings" window completely redesigned with new options and improved
OS X support (shows up in Preferences instead of in a different pull
down menu)

-Pre-Process Settings window redesigned to get rid of that weird
tree-tab thing Jeff had going on and substituted in instead legitimate
Java tabs for improved ease of use.

-Improvements to PropertiesUtil to make it easier for developer to set
and get settings.

-Translated sentences "priority queue" completed (when a sentence is
changed or swapped with a translation, it will be translated again
immediately after the current sentence is done being processed)